requirements.txt
aiogram==3.13.1
google-generativeai==0.8.3
python-dotenv==1.0.1